+++ src/components/TreeMap/ITreeMapState.ts
@@ -1,6 +1,6 @@
+import { IBreadcrumbItem } from "../Breadcrumb";
 import { CustomHierarchyRectangularNode } from "./TreeMap";
 export interface ITreeMapState<TreeMapInputData> {
-  scopedNodes: Array<CustomHierarchyRectangularNode<TreeMapInputData>>;
   data: TreeMapInputData;
   width: number;
   height: number;
@@ -9,9 +9,8 @@ export interface ITreeMapState<TreeMapInputData> {
   xScaleFunction: any;
   yScaleFunction: any;
   zoomEnabled: boolean;
-  breadCrumbItems: any;
+  breadcrumbItems: IBreadcrumbItem[];
   selectedId: number;
   selectedNode: CustomHierarchyRectangularNode<TreeMapInputData>;
   totalNodes: number;
-  selectedNodeTotalNodes: number;
 }